requirements.txt
langchain==1.1.0
langchain-openai==1.1.0
langchain-community==0.4.1
langchain-classic==1.0.0
python-dotenv==1.2.1
numpy<2.0.0